fix possible undefined cases in text channel creation logic

+++ src/commands/voice/channelLeave.ts
@@ -4,9 +4,15 @@ import { updateTextChannelVisibility } from "./updateTextChannelVisibilty.js";
 
 export async function voiceChannelLeave(oldState: VoiceState) {
     const leftVoiceChannel = oldState.channel
+    if (leftVoiceChannel === null || oldState.member === null) {
+        return
+    }
     const oldVoiceMembers = leftVoiceChannel.members
     const textChannelId = textToVoiceId.get(leftVoiceChannel.id)
+    if (textChannelId === undefined) { return }
+
     const textChannel = await oldState.guild.channels.fetch(textChannelId)
+    if (textChannel === null) { return }
 
     if (oldVoiceMembers.size <= 0) { //last member left voice channel
         textChannel.delete();
